+++ src/modules/app_center/views/master_analysis/promptlab/components/PromptlabPanel.ts
@@ -69,6 +69,7 @@ import {
   restoreInput,
   toggleConsoleMode,
   copyPrompt,
+  copySeoKeywords,
   handoffListingPromptToDeepChat,
   clearInputs,
   selectAllReportSections,
@@ -769,6 +770,10 @@ const promptlabPanelBehavior: PromptlabPanelBehavior = {
 
   copyPrompt() {
     copyPrompt();
+  },
+
+  async copySeoKeywords() {
+    await copySeoKeywords(this as unknown as PromptlabAlpineContext);
   },
 
   async handoffListingPromptToDeepChat() {